Moved firebase files to /db/firebase for consistency and cleaner structure
