+++ classes/verification.js
@@ -1,6 +1,6 @@
 const { GuildMember, Guild } = require('discord.js');
 const discordServices = require('../discord-services');
-const firebaseServices = require('../firebase-services/firebase-services');
+const firebaseServices = require('../db/firebase/firebase-services');
 const BotGuildModel = require('./bot-guild');
 
 /**
