+++ commands/a_boothing/start-boothing.js
@@ -1,7 +1,7 @@
 // Discord.js commando requirements
 const PermissionCommand = require('../../classes/permission-command');
-const firebaseServices = require('../../firebase-services/firebase-services');
-const firebaseBoothing = require('../../firebase-services/firebase-services-boothing')
+const firebaseServices = require('../../db/firebase/firebase-services');
+const firebaseBoothing = require('../../db/firebase/firebase-services-boothing')
 const discordServices = require('../../discord-services');
 const Discord = require('discord.js');
 const BotGuildModel = require('../../classes/bot-guild');
