+++ commands/verification/check-member.js
@@ -1,6 +1,6 @@
 // Discord.js commando requirements
 const PermissionCommand = require('../../classes/permission-command');
-const firebaseServices = require('../../firebase-services/firebase-services');
+const firebaseServices = require('../../db/firebase/firebase-services');
 const discordServices = require('../../discord-services');
 const Discord = require('discord.js');
 const Prompt = require('../../classes/prompt');
